requirements.txt
Django==3.2
environs==9.5.0
Pillow==9.4.0
celery==5.2.*
django-celery-results==2.2.*
redis==4.5.*
qrcode==7.4.*
yookassa==2.3.5
django-celery-beat==2.5.*
django-phonenumber-field==7.0.2
phonenumbers==8.13.4